% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_count.R
\name{empty_count}
\alias{empty_count}
\title{empty_count}
\usage{
empty_count(dataframe)
}
\arguments{
\item{dataframe}{Dataframe representing metadata, 'Section' column as factor}
}
\value{
Returns a long dataframe with 3 columns: Empty (No, Yes),
Table (table name), N_Variables (count).
}
\description{
Internal Function: Called within the map_compare.R function. \cr \cr
It reads in the metadata dataframe and counts how many of the variables have
empty descriptions.
}
\keyword{internal}
