% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{emptydesc_count}
\alias{emptydesc_count}
\title{emptydesc_count}
\usage{
emptydesc_count(table_df, table_colname)
}
\arguments{
\item{table_df}{Output from json_table_to_df.R}

\item{table_colname}{Name of the table in column of the output dataframe}
}
\value{
A simpler dataframe with Y/N empty counts for variables in the table.
}
\description{
Internal Function: This function is called within the metadata_map function. \cr \cr
It reads in a data frame that summarises one table of the dataset. \cr \cr
It counts missing variable descriptions, based on specified criteria.
}
\keyword{internal}
