% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{end_plot}
\alias{end_plot}
\title{Internal: end_plot}
\usage{
end_plot(df, table_name, ref_table)
}
\arguments{
\item{df}{The Output dataframe with all the domain categorisations}

\item{table_name}{Table name}

\item{ref_table}{Domain code reference table (domains mapped to integers)}
}
\value{
It returns a ggplot
}
\description{
This function is called within the metadata_map function. \cr \cr
A summary plot is created that includes the domain code reference table and
counts of domain code categorisations \cr \cr
}
\seealso{
Other metadata_map_internal: 
\code{\link{data_load}()},
\code{\link{empty_count}()},
\code{\link{empty_plot}()},
\code{\link{output_copy}()},
\code{\link{user_categorisation}()},
\code{\link{user_categorisation_loop}()}
}
\concept{metadata_map_internal}
\keyword{internal}
\keyword{internal}
