% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapmetadata-package.R
\docType{package}
\name{mapmetadata-package}
\alias{mapmetadata}
\alias{mapmetadata-package}
\title{mapmetadata: Map health metadata onto predefined research domains}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Prior to gaining full access to health datasets, explore publicly available metadata and map metadata onto predefined research domains. This package uses structural metadata files downloaded from the Health Data Research Gateway (https://healthdatagateway.org/en). In theory, any metadata file with the same structure as the files downloaded from this gateway can be used with this package, but the package has been developed and tested on metadata files from this gateway only.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/mapmetadata/}
  \item Report bugs at \url{https://github.com/ropensci/mapmetadata/issues}
}

}
\author{
\strong{Maintainer}: Rachael Stickland \email{rstickland@turing.ac.uk} (\href{https://orcid.org/0000-0003-3398-4272}{ORCID})

Other contributors:
\itemize{
  \item Batool Almarzouq (\href{https://orcid.org/0000-0002-3905-2751}{ORCID}) [contributor]
  \item Mahwish Mohammad (\href{https://orcid.org/0009-0004-5295-0726}{ORCID}) [contributor]
  \item Daniel Delbarre (\href{https://orcid.org/0000-0003-4633-4252}{ORCID}) [contributor]
  \item Nida Ziauddeen (\href{https://orcid.org/0000-0002-8964-5029}{ORCID}) [contributor]
  \item Zoë Turner (Zoë reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/674) [reviewer]
  \item Yohann Mansiaux (Yohann reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/674) [reviewer]
}

}
\keyword{internal}
