% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package_data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Internal: CSV metadata File}
\format{
Nested lists
}
\source{
https://healthdatagateway.org/en/dataset/360
}
\usage{
data(metadata)
}
\description{
Example metadata for a health dataset, to demo metadata_map \cr \cr
This data was created with these steps:
\enumerate{
\item Go to https://healthdatagateway.org
\item Navigate to the dataset of interest, select 'Download data' and
download the Structural Metadata file
\item Shorten name of downloaded file e.g. 360_NCCHD_Metadata.csv
\item \code{metadata <- read.csv(system.file('inputs',
  '360_NCCHD_Metadata.csv', package = 'mapmetadata'))}
\item \code{usethis::use_data(metadata)}
}
}
\keyword{internal}
