% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{outputs_join}
\alias{outputs_join}
\title{outputs_join}
\usage{
outputs_join(session_1, session_2)
}
\arguments{
\item{session_1}{Dataframe from session 1}

\item{session_2}{Dataframe from session 2}
}
\value{
Dataframe with information from session 1 and 2, joined on column DataElement.
}
\description{
Internal Function: This function is called within the map_compare.R function. \cr \cr
Joins output dataframes from two sessions, on the column DataElement.
}
\keyword{internal}
