% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interactions.R
\name{user_categorisation}
\alias{user_categorisation}
\title{Internal: user_categorisation}
\usage{
user_categorisation(var, desc, type, domain_code_max)
}
\arguments{
\item{var}{Name of the variable}

\item{desc}{Description of the variable}

\item{type}{Data type of the variable}

\item{domain_code_max}{Max code in the domain list (0-3 auto included,
then N included via domain_file)}
}
\value{
It returns a list containing the decision and decision note
}
\description{
Internal Function: Called within the metadata_map function. \cr \cr
It displays data properties to the user and requests a categorisation into
a domain. \cr \cr
An optional note can be included with the categorisation.
}
\seealso{
Other metadata_map_internal: 
\code{\link{data_load}()},
\code{\link{empty_count}()},
\code{\link{empty_plot}()},
\code{\link{end_plot}()},
\code{\link{output_copy}()},
\code{\link{user_categorisation_loop}()}

Other map_compare_internal: 
\code{\link{consensus_on_mismatch}()},
\code{\link{valid_comparison}()}
}
\concept{map_compare_internal}
\concept{metadata_map_internal}
\keyword{internal}
\keyword{internal}
