% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{valid_comparison}
\alias{valid_comparison}
\title{valid_comparison}
\usage{
valid_comparison(input_1, input_2, severity, severity_text)
}
\arguments{
\item{input_1}{Input 1}

\item{input_2}{Input 2}

\item{severity}{Level of severity. Only 'danger' or 'warning'}

\item{severity_text}{The text to print if inputs are not equal.}
}
\value{
Returns nothing if inputs are equal. If inputs are not equal,
returns variable text depending on level of severity.
}
\description{
Internal Function: Called within the map_compare.R function. \cr \cr
It reads two inputs to see if they are equal. \cr \cr
Warning severity: if unequal inputs, gives warning but continues. \cr \cr
Danger severity: if unequal inputs, exits with error message. \cr \cr
}
\seealso{
Other map_compare_internal: 
\code{\link{consensus_on_mismatch}()},
\code{\link{user_categorisation}()}
}
\concept{map_compare_internal}
\keyword{internal}
\keyword{internal}
