% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_collect.R
\name{data_load}
\alias{data_load}
\title{data_load}
\usage{
data_load(json_file, domain_file, look_up_file)
}
\arguments{
\item{json_file}{As defined in metadata_map}

\item{domain_file}{As defined in metadata_map}

\item{look_up_file}{As defined in metadata_map}
}
\value{
A list of 5: all inputs needed for the metadata_map function to run.
}
\description{
Internal Function: This function is called within the metadata_map function. \cr \cr
It collects the inputs needed for the metadata_map function (defaults or user inputs)
If some inputs are NULL, it loads the default inputs. If defaults not available, it prints error for the user.
If inputs are not NULL, it loads the user-specified inputs.
}
\keyword{internal}
