% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_collect.R
\name{data_load}
\alias{data_load}
\title{data_load}
\usage{
data_load(metadata_file, domain_file, look_up_file, quiet = FALSE)
}
\arguments{
\item{metadata_file}{As defined in metadata_map}

\item{domain_file}{As defined in metadata_map}

\item{look_up_file}{As defined in metadata_map}

\item{quiet}{Default is FALSE. Change to TRUE to quiet the cli_alert_info
and cli_alert_success messages.}
}
\value{
A list of 6: all inputs needed for the metadata_map function to run.
}
\description{
Internal Function: Called within the metadata_map function. \cr \cr
Collects inputs needed for metadata_map function (defaults or user inputs)
If some inputs are NULL, it loads the default inputs. If defaults not
available, it prints error for the user.
If inputs are not NULL, it loads the user-specified inputs.
}
\keyword{internal}
\keyword{internal}
