% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{empty_plot}
\alias{empty_plot}
\title{Internal: empty_plot}
\usage{
empty_plot(dataframe, bar_title)
}
\arguments{
\item{dataframe}{Dataframe to plot. Output of empty_count.R function.}

\item{bar_title}{Title for the plot.}
}
\value{
A plotly object
}
\description{
This function is called within the metadata_map function. \cr \cr
It plots  \cr \cr
}
\keyword{internal}
