% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package_data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Internal: CSV metadata File}
\format{
Nested lists
}
\source{
https://modelcatalogue.cs.ox.ac.uk/hdruk_live/#/catalogue/dataModel/16920b16-e24c-49f9-b4df-3dc85779822b/dataClasses
}
\usage{
data(metadata)
}
\description{
Example metadata for a health dataset, to demo the function metadata_map \cr \cr
This data was created with these steps:
\enumerate{
  \item Go to https://healthdatagateway.org
  \item Navigate to the dataset of interest, select 'Download data' and download the Structural Metadata file
  \item \code{metadata <- read.csv("inst/inputs/360_National Community Child Health Database (NCCHD)_Structural_Metadata.csv")}
  \item \code{usethis::use_data(metadata)}
}
}
\keyword{internal}
