% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interactions.R
\name{user_categorisation_loop}
\alias{user_categorisation_loop}
\title{Internal: user_categorisation_loop}
\usage{
user_categorisation_loop(
  start_v,
  end_v,
  table_df,
  df_prev_exist,
  df_prev,
  lookup,
  df_plots,
  output_df,
  quiet = FALSE
)
}
\arguments{
\item{start_v}{Index of data element to start}

\item{end_v}{Index of data element to end}

\item{table_df}{Dataframe with table info, extracted from metadata file}

\item{df_prev_exist}{Boolean to indicate if previous dataframes exists}

\item{df_prev}{Previous dataframes to copy from (or NULL)}

\item{lookup}{The lookup table to enable auto categorisations}

\item{df_plots}{Output from ref_plot, to indicate mac domain code allowed}

\item{output_df}{Empty output dataframe, to fill}

\item{quiet}{Default is FALSE. Change to TRUE to quiet the cli_alert_info
and cli_alert_success messages.}
}
\value{
An output dataframe containing info about the table, data elements
and categorisations
}
\description{
Internal Function: Called within the metadata_map function. \cr \cr
Given a specific table and a number of data elements to search, it checks for
3 different sources of domain categorisation: \cr \cr
1 - If data elements match look-up table, auto categorise \cr \cr
2 - If data elements match to previous table output, copy them \cr \cr
3 - If no match for 1 or 2, data elements are categorised by the user \cr \cr
}
\keyword{internal}
\keyword{internal}
